buildin.rsのSTEP数がかなり増えたので、分割した
